Rota de update só permite alterações em moedas ficticias

+++ currency-conversion/currency-conversion.web/Controllers/CurrencyController.cs
@@ -52,10 +52,14 @@ namespace currency_conversion.web.Controllers
         }
 
         [HttpPost("{code}")]
-        public async Task<IActionResult> InsertAsync([RegularExpression("^[0-9a-zA-Z]{3,10}$", ErrorMessage = "Field must contain only 0-9 a-z A-Z characters, 3 to 10 characters")] string code)
+        public async Task<IActionResult> Insert([RegularExpression("^[0-9a-zA-Z]{3,10}$", ErrorMessage = "Field must contain only 0-9 a-z A-Z characters, 3 to 10 characters")] string code)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var currencyToAdd = await _currencyFetch.FetchCurrencyAsync(code);
-            if (currencyToAdd == null) return BadRequest("Currency not supported");
+            if (currencyToAdd == null) return BadRequest("Currency not supported by exchange provider");
             var created = _currencyRepository.Create(currencyToAdd);
             if (!created) return BadRequest("Currency could not be added: Code already exists");
             return Ok("Currency added");
@@ -70,13 +74,17 @@ namespace currency_conversion.web.Controllers
             }
             var _mappedInputCurrency = _mapper.Map<Currency>(currencyDTO);
             var updated = _currencyRepository.Update(_mappedInputCurrency);
-            if (!updated) return NotFound("Currency not found: " + currencyDTO.Code);
+            if (!updated) return NotFound("Currency not found: " + currencyDTO.Code + ", or not eligible for update");
             return Ok("Currency updated");
         }
 
-        [HttpDelete(Name = "delete_custom")]
-        public IActionResult DeleteCustom([RegularExpression("^[0-9a-zA-Z]{3,10}$", ErrorMessage = "Field must contain only 0-9 a-z A-Z characters, 3 to 10 characters")]string code)
+        [HttpDelete(Name = "delete")]
+        public IActionResult Delete([RegularExpression("^[0-9a-zA-Z]{3,10}$", ErrorMessage = "Field must contain only 0-9 a-z A-Z characters, 3 to 10 characters")]string code)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var deleted = _currencyRepository.Delete(code);
             if (!deleted) return NotFound("Currency not found: " + code);
             return Ok("Currency deleted");
